pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcpagent"
version = "0.1.0"
description = "Universal AI agent with MCP tool integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "mcp>=1.20",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
mcpagent = "mcpagent.__main__:main_entry"

[tool.hatch.build.targets.wheel]
packages = ["src/mcpagent"]
